Move general geometry methods to new package #545
